requirements.txt
Flask==2.0.3
mysql-connector-python==8.0.26
Flask-Cors==3.0.10
python-dotenv==0.19.2